pyproject.toml
[tool.poetry]
name = "fastapi-auth"
version = "0.1.0"
description = ""
authors = ["Raj Singh <raj.tanwar@techdigitem.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
motor = "^3.3.2"
httpx = "^0.26.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.13"
black = "^23.12.1"
mypy = "^1.8.0"
pytest = "^7.4.4"
coverage = "^7.4.0"
mongomock-motor = "^0.0.26"
pytest-httpx = "^0.28.0"
pytest-asyncio = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.ruff]
select = ["E", "F", "I"]
fixable = ["ALL"]
exclude = [".git", ".mypy_cache", ".ruff_cache"]
line-length = 88

[tool.mypy]
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
disallow_untyped_decorators = false
ignore_missing_imports = true
implicit_reexport = true
plugins = "pydantic.mypy"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
